Fix crash on DPU w/ libdocahal_trace LD_PRELOADed
nvp4c currently crashes during execution of global fini/destructor
routines if libdocahal_trace was LD_PRELOADED because:

1) libdocahal_trace's global init/constructor runs, which:
   a) Calls dlsym() to determine the address of a symbol in libdocahal.
   b) dlsym() calls calloc() to allocate heap storage.
   c) dlsym() stores the resultant pointer in TLS (Thread Local
      Storage).
   libgc does not treat TLS data as a "root" and so is not aware that
   this memory block is still referenced.

2) The cstring cache's global constructor runs, which allocates some
   memory via new. libgc ends up freeing the memory block allocated in
   (1) above since it appears unreferenced. Consequently, libgc returns
   the same block/pointer for use by cstring.

3) During compiler operation, a function is called within libdocahal_trace,
   which internally calls dlsym() again. This causes libdl to write into
   the pointer it shares with cstring. This causes memory corruption of a
   cstring cache std::set internal data structure.

4) At the end of nvp4c operation, global destructors are executed. This
   includes the destructor for the cstring cache. It crashes with a
   segfault because some internal pointer now has value 4G (0x100000000)
   due to point (3) above.

There are various ways to solve this, none of the entirely perfect. This
commit implements two of them, and enables only one of them. The other is
implemented and left in the code just in case we need to switch technique
later, to avoid having to re-implement the technique at that time.

Solution 1 (enabled):

(Suggested by Chris Dodd, implemented by me)

Register the TLS data as an additional root for libgc. This allows it to
see that the data allocated by libdl is still referenced, so it is not
freed, and so (2) above does not happen.

The current implementation has some deficiences in the face of a
multi-threaded program, which is currently not relevant in nvp4c.

Solution 2 (disabled):

Detect any allocation made by libdl, and implement it in a way that tells
libgc never to collect the allocation.

The current implementation has some deficiencies:
- Is implemented solely for calloc, not for malloc or realloc. This is
  currently fine, since libdl only calls calloc; confirmed with objdump.
- Is applied to every allocation made by libdl. This  is currently fine
  since IIRC dlsym only makes this one allocation.
- Doesn't address potential TLS use by any other code.

Both solutions passed _many_ hundreds of test runs, whereas without this
fix, the issue could easily be reproduced in just 3-5 invocations of
nvp4c.