expression: Actively fail the compilation when key value is over bitWidth
The front-end was shrinking the value when the key value is over the target
bitwidth with warnings. However, this may lead to the duplicate const entry key
values at the later stage. Instead of warnings, actively fails the compilation
as this will lead to the problematic compilation output.

While we do not support singed key value, add the error check for signed key
value for unsigned key.