handleOverflow: Bypass cast case for the overflow check
In the current implementation of the frontend code, (value & mask) != value may
be valid as the casting shrinks the value by the updated target bitwidth size
(actual usage). The front-end code assumes this case and thus issuing warnings
instead of errors to cover following cases:
- Actual overflow
- False overflow warning due to the updated target bitwidth size based on the
  usage, which in fact is valid as the casting will disregard first few bits.
This commit updates the front-end code such that it validates the overflow for
non-cast constant expression.