Use NV mirrors for git repos
p4c's build process now downloads certain dependencies from a git server
during the build, rather than requiring the code to have been previously
downloaded/cloned by the developer prior to launching the build. Modify
the URLs of relevant git servers to point at NV servers instead of
external servers. This (a) complies with NV policy of not relying on
external servers, (b) hopefully is more reliable and efficient, since
network traffic is more local.

Note that the inja URL is invalid; we do not yet have a mirror of this
repo. However, (a) this repo isn't actually used by the build process at
present, (b) if we do ever mirror the repo, this should be the URL that
we mirror it to.