Disable or comment non-git downloads from 3rd-party servers
Use invalid URLs for any unused downloads from 3rd-party servers, to
ensure that if the relevant code is ever triggered, it will fail, so we
will become aware of the need to mirror the relevant files.

Add FIXME comments for any used downloads from 3rd-party servers; we
need to mirror these files on some internal server to comply with NV
policy. However, the files aren't mirrored yet, so we can't actually
replace the URLs yet.